app_auth/flask/requirements.txt
flask
mysql-connector-python==8.0.5
requests
pymysql